LayoutTests/http/tests/referrer-policy/no-referrer/about-blank-inherits-opener-referrer-policy-expected.txt

Tests that about:blank popup inherits the opener's no-referrer referrer policy.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is ""
PASS successfullyParsed is true

TEST COMPLETE
